Improved error management
Moving certain blocks of code inside the try catch statements in command files to ensure that if something goes wrong it is catched and an error message is sent to the user as well as the console. Also moving out deferred responses from the try catch statements to prevent double errors if the interaction is expired. This commit also adds error messages for missing environment variables to make any problems caused by them less confusing. Now if anywhere any environment variables are missing the program will throw error and stop executing there.