AM4 Bot v5.0.1
Improving some of the utility functions and redoing the settings handling for the AM4 rest client. In addition, improving typings and switching from node-fetch to undici. V5.0.0 final release was supposed to be the last commit, but decided to finish the last unfinished improvements that I did not have time to finish before 28th of January. This will 100% be the last commit and no more fixes and/or improvements will be done to the code anymore.