Archiving the source code
Updating the README.md to inform that the source code is archived and therefore opening new discussions will not be possible (not like any were opened at any point). Also couldn't resist to make the last type improvements to the code, so I have included many improvements and fixes to the typings.