Improvements
Improved typings and improved usage of typeguards instead of overwriting types.

Improved the leftover seat configuration calculation. In addition, adding more methods to the Plane namespace (previously a class with static methods).

Fixed some minor bugs in the code that were found. In addition, removed any Telegram and/or Discord invite URLs from the code and made them environment variables. This way it is simpler to manage and change them on the fly.